Add some features to common requests, add simple test, CI and docs deploy
The ideas is to use tests to document various features, as
embedded-hal-mock is so nice with helping. This also tests the driver
simultaneously.

Deploy the docs for now to gh pages, this is mostly for me to work on
them and see what is going on on a regular basis.